requirements.txt
asgiref==3.8.1
captcha==0.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-axes==7.0.2
django-geoip2-extras==4.1
django-otp==1.6.0
django-recaptcha==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
geoip2==3.0.0
idna==3.10
maxminddb==2.6.3
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
qrcode==8.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
